build supporting arrays in dedicated utils
